refactor: move StatusColorsExt to theme package
